Added potential code for implementing scale.
Code is commented out in Garden.tsx (starting on line 83)

Includes state for the garden size,
a function that returns the percentage of the space
that a prop will take up in the garden,
and a textbox that will update the garden size upon entry.
Updating the garden size will change the size of the "box"
that contains the prop images.
It will do so by changing the width (and height) of the box
to be equal to the percentage calculated in the
getProptoGardenPercentage function.